requirements_simple.txt
# Core dependencies (without TA-Lib)
python-binance>=1.0.19
websockets>=11.0.3
aiohttp>=3.8.5

# Data processing and analysis
pandas>=2.0.3
numpy>=1.24.3
scipy>=1.11.1

# Security and encryption
cryptography>=41.0.3

# Configuration and utilities
pydantic>=2.1.1
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Logging and monitoring
structlog>=23.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-mock>=3.11.1